Extract Court type
RISDEV-8322
