+++ frontend/src/components/DocumentUnitSearch.vue
@@ -9,7 +9,8 @@ import InfoModal from "@/components/InfoModal.vue"
 import Pagination, { Page } from "@/components/Pagination.vue"
 import { useInternalUser } from "@/composables/useInternalUser"
 import { Query } from "@/composables/useQueryFromRoute"
-import { Court, DocumentUnitSearchParameter } from "@/domain/documentUnit"
+import { Court } from "@/domain/court"
+import { DocumentUnitSearchParameter } from "@/domain/documentUnit"
 import DocumentUnitListEntry from "@/domain/documentUnitListEntry"
 import errorMessages from "@/i18n/errors.json"
 import comboboxItemService from "@/services/comboboxItemService"
