+++ frontend/test/components/activeCitations.spec.ts
@@ -9,9 +9,9 @@ import { createRouter, createWebHistory } from "vue-router"
 import ActiveCitations from "@/components/ActiveCitations.vue"
 import ActiveCitation from "@/domain/activeCitation"
 import { CitationType } from "@/domain/citationType"
+import { Court } from "@/domain/court"
 import { Decision } from "@/domain/decision"
 import { DocumentType } from "@/domain/documentType"
-import { Court } from "@/domain/documentUnit"
 
 import documentUnitService from "@/services/documentUnitService"
 import featureToggleService from "@/services/featureToggleService"
