+++ frontend/test/components/comboboxInput.spec.ts
@@ -4,7 +4,7 @@ import { http, HttpResponse } from "msw"
 import { setupServer } from "msw/node"
 import ComboboxInput from "@/components/ComboboxInput.vue"
 import { ComboboxAttributes, ComboboxItem } from "@/components/input/types"
-import { Court } from "@/domain/documentUnit"
+import { Court } from "@/domain/court"
 import comboboxItemService from "@/services/comboboxItemService"
 
 const items = [
