+++ frontend/test/components/documentUnit/documentUnitSearch.spec.ts
@@ -7,7 +7,7 @@ import { setupServer } from "msw/node"
 import InputText from "primevue/inputtext"
 import { createRouter, createWebHistory } from "vue-router"
 import DocumentUnitSearch from "@/components/DocumentUnitSearch.vue"
-import { Court } from "@/domain/documentUnit"
+import { Court } from "@/domain/court"
 import DocumentUnitListEntry from "@/domain/documentUnitListEntry"
 import authService from "@/services/authService"
 import documentUnitService from "@/services/documentUnitService"
