+++ frontend/test/components/ensuingDecisions.spec.ts
@@ -7,10 +7,10 @@ import { setupServer } from "msw/node"
 import InputText from "primevue/inputtext"
 import { createRouter, createWebHistory } from "vue-router"
 import EnsuingDecisions from "@/components/EnsuingDecisions.vue"
+import { Court } from "@/domain/court"
 import { Decision } from "@/domain/decision"
 import { DocumentType } from "@/domain/documentType"
 
-import { Court } from "@/domain/documentUnit"
 import EnsuingDecision from "@/domain/ensuingDecision"
 import documentUnitService from "@/services/documentUnitService"
 import featureToggleService from "@/services/featureToggleService"
