+++ frontend/test/components/previousDecisions.spec.ts
@@ -7,9 +7,9 @@ import { setupServer } from "msw/node"
 import InputText from "primevue/inputtext"
 import { createRouter, createWebHistory } from "vue-router"
 import PreviousDecisions from "@/components/PreviousDecisions.vue"
+import { Court } from "@/domain/court"
 import { Decision } from "@/domain/decision"
 import { DocumentType } from "@/domain/documentType"
-import { Court } from "@/domain/documentUnit"
 
 import PreviousDecision from "@/domain/previousDecision"
 import documentUnitService from "@/services/documentUnitService"
